Move ui modules and update paths
